requirements.txt
Flask
Flask-JWT-Extended
Flask-Bcrypt
Flask-Cors
Flask-Mail
Flask-PyMongo
faiss-cpu
pdfplumber
sentence-transformers
Flask-Dance
authlib
python-dotenv
